golioth_coap_client: verify response token matches
Previously, we used a bool flag to indicate when a request
had received a response, but this was buggy in cases where we get
other kinds of responses (like observations) before getting
the response to the request we were expecting, which causes
retransmits and things generally getting out of sync in the
main IO loop.

This fixes the bug by saving the token from the request and
comparing it to the received token.

Signed-off-by: Nick Miller <nick@golioth.io>