golioth_coap_client: process io if no requests
If there are no requests to process, we still need to process
libcoap IO in case there are observations.

Also, lowered the queue timeout value to 1 s. This represents
how often we poll for observations.

There does not appear to be any better way to check for
observations than polling. A second option would be to orient
the client around processing IO instead of the request queue,
which would make us responsive to observations, but would
require polling of the request queue. It seems that as a
client, we want to be responsive to processing requests,
at the expense of slower polling of async observations, which is
why this second option was not chosen.

Signed-off-by: Nick Miller <nick@golioth.io>