Kconfig: default keepalive 9000 ms
The default for keepalive should be enabled.
It is basically guaranteed that NAT issues will occur if
set to 0 (disabled), which prevents CoAP observations
from being delivered.

The value of 9 seconds was copied from the Zephyr SDK and
was tested to work over long periods of device operation.
It's likely this can be relaxed to 30 seconds, but this
has not been tested.

Signed-off-by: Nick Miller <nick@golioth.io>