coap_client: add initial empty request to queue
The initial empty request is intended as a way to
verify we are connected to Golioth when the session
is first established.

Previously, we bypassed the request queue and sent the
empty request directly in the client task.
However, doing it this way does not trigger the logic
that detects connection events, so the user does not
get a callback when the empty response is received
from Golioth.

To fix, we simply put the initial empty request into
the request queue, so it can be processed the same way
as all other requests, so that connection events can
be properly detected.

Signed-off-by: Nick Miller <nick@golioth.io>