golioth_coap_client: copy user's path
The prior assumption was that all path strings are
const static string literals, such that the CoAP task
can access the string at any point in time without
worrying about memory lifetime.

However, this is not always the case, and it’s not obvious
that this is the assumption being made (unless one reads
the SDK comments and code carefully). It can be surprising
to user's if they are using a dynamically generated path.

Instead, copy the user's path and store it in the request.

Signed-off-by: Nick Miller <nick@golioth.io>