Add response status to all client callbacks
In general, if there is a request timeout, or a response
contains a non-success code (e.g. 4.XX), the user is never
made aware of it. This can cause silent failures in OTA,
where the fw_update code thinks it has received a valid
block when in fact it hasn't.

This commit updates the callback APIs, but does not include
any of the error handling.

Signed-off-by: Nick Miller <nick@golioth.io>