Add timeout param to synchronous APIs (#16)
* sync APIs have timeout param

Signed-off-by: Nick Miller <nick@golioth.io>

* New file golioth_time.h

Both the SDK and the examples need the millis/micros/delay_ms
functions, so it makes sense to include them in the SDK as
a public header file.

Signed-off-by: Nick Miller <nick@golioth.io>

* coap_client: age out sync requests

For synchronous requests, it's not always desirable from a
user's perspective to wait forever for a response.
Is a session is established, there is a 10-second default timeout
for each request, which will cause the sync function to return
if reached. However, if there is no session (e.g. no Internet
connection), then the user will be left waiting potentially
forever for the request to complete, which is often not
desirable.

The _sync APIs now take a timeout_s param, so user's can specify
a custom request timeout. Internally, this is stored as a request
ageout parameter, which is a timestamp for when the request is
no longer valid. If the client task pulls a request out of the
queue that is past the ageout value, it will drop it. Also, if
a response is received for a request that has aged out, the
response will dropped (request callbacks not called).

The client task will wait until either the 10-second default
timeout is reached, or the ageout time, whichever is sooner,
at which point the request is completed and the _sync functions
will be unblocked.

Signed-off-by: Nick Miller <nick@golioth.io>

* verify.py: skip ota test if BIT fails

Signed-off-by: Nick Miller <nick@golioth.io>

* fw_update: normal logging of downloaded blocks

Signed-off-by: Nick Miller <nick@golioth.io>

* coap_client: more debug logs for block GET

Signed-off-by: Nick Miller <nick@golioth.io>