fw_update: ignore manifest if download in progress
Fixes an issue observed in CI where an OTA was in progress on
a device, and at the same time a new firmware image was uploaded
to Golioth from a separate CI job.

This caused the image size in the manifest to change while
the download was in progress, resulting in the final image size
check to fail.

Signed-off-by: Nick Miller <nick@golioth.io>