add callback to all async set/delete functions
Previously, there was no way to be notified when a set or
delete async call received a response from the server,
and no way to know if there was an error.

This commit changes the existing set and delete async
functions to also pass optional callback function and
void* callback arg.

This is immediately useful in the test example
(test_lightdb_set_get_async), to avoid waiting "in the dark"
for an async set call to receive a response, and it seems
like it will be generally useful for SDK users, at the expense
of a couple of extra parameters at the call site.

Signed-off-by: Nick Miller <nick@golioth.io>