test: insert delay after sync set
Previously, the test_lightdb_set_get_sync function would
fail intermittently, due to a race condition between the
server writing the data to the database, and the SDK trying to
read it back. In the failing case, the data would be read before
the prior write had completely finished writing to the database.

To worked around this behavior of the server
("eventual consistency"), we insert a small delay before trying
to read the data back.

Signed-off-by: Nick Miller <nick@golioth.io>