test: add delay in async test
In the test_lightdb_set_get_async test, we need to insert
a delay after setting the value, to give the server time
to write the value to the database.

We already had this delay in test_light_set_get_sync, but
forgot to add it to this test, and it causes CI to fail
about 10% of the time.

Signed-off-by: Nick Miller <nick@golioth.io>