verify.py: Increase timeout for OTA start
Sometimes CI will timeout waiting for to see "State = Downloading"
from the device. This is happening because the initial empty
keepalive for the CoAP session sometimes doesn't get a response
from the server for a very long time (5+ seconds).

Doubling the timeout for that step should fix the flaky test.

Signed-off-by: Nick Miller <nick@golioth.io>