lightdb: strtol with NULL-terminated string
In golioth_payload_as_int/float, we are using strtol and strtof
to covert from string to int/float. However, since the payload
is not null-terminated, these can return the wrong value if
the bytes after the payload happen to be digits.

To fix, we need to copy the payload to temporary buffer that
is guaranteed to be NULL-terminated.

Signed-off-by: Nick Miller <nick@golioth.io>