test: start_ota should run in Unity context
The prior code would crash in the start_ota shell command
if any of the Unity TEST_ASSERT_* assertions failed.
This was happening occassionally in CI if the initial
EMPTY request was not responded to within 10 seconds
(which is a separate issue).

I think the reason the crash occurs is because those
assert macros only work in the context of a Unity session
(indicated by UNITY_BEGIN and UNITY_END macros).

To fix, the start_ota command now executes the
connect functions in the context of a Unity test
session, similar to built_in_test.

Signed-off-by: Nick Miller <nick@golioth.io>