coap_client: verbose error logs if nack received
Previously, if the PSK was not valid, the only error you'd
see is "Timeout" and "Received nack reason: 3", which is not
useful to most people.

Since this is a common error, we need better logging.
In this case, nack reason 3 is COAP_NACK_TLS_FAILED,
so print a more human-readable string and a
hint that the PSK-ID or PSK might be wrong.

Signed-off-by: Nick Miller <nick@golioth.io>