.github: move self-hosted env to runner_env.sh
The self-hosted runner workflows will now assume there is
a file in $HOME/runner_env.sh which defines the environment
variable for the ESP32S3 serial port.

This is simpler than trying to set the environment in the
GitHub Actions service.

Also added comments to more clearly spell out the assumptions
made about the self-hosted runner, and the files that must
be present.

Signed-off-by: Nick Miller <nick@golioth.io>