magtag_demo: read golioth credentials from NVS
The README states that golioth/psk-id and golioth/psk
should be set in the settings shell, but app_main
is still expecting them to be set via sdkconfig.

Remove support for PSK via sdkconfig and read from
NVS instead, to be consistent with README and other
examples like golioth_basics.

As part of this, delete Kconfig.projbuild to prevent
setting default values for the now unsupported Kconfig
items.

Signed-off-by: Nick Miller <nick@golioth.io>